# Improve complex airy asymptotic expansions