--- a/Project.toml
+++ b/Project.toml
@@ -1,9 +1,13 @@
 name = "Bessels"
 uuid = "0e736298-9ec6-45e8-9647-e4fc86a2fe38"
-version = "0.3-DEV"
+version = "0.3.0-DEV"
+
+[deps]
+SIMDMath = "5443be0b-e40a-4f70-a07e-dcd652efc383"
 
 [compat]
 julia = "1.6"
+SIMDMath = "0.2.1"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"