--- a/src/Bessels.jl
+++ b/src/Bessels.jl
@@ -1,5 +1,7 @@
 module Bessels
 
+using SIMDMath
+
 export besselj0
 export besselj1
 export besselj