The previous large argument expansions were kind of a mess. They were also much slower than they needed be. By unrolling the expansions to a certain degree we can avoid any complex divisions.

```julia
# latest tagged version
julia> @benchmark Bessels.airyai_large_argument(x) setup=(x=(8.3 + rand()*2)*cispi(2*rand()))
BenchmarkTools.Trial: 10000 samples with 142 evaluations.
 Range (min … max):  705.690 ns …   2.018 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     834.965 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   967.171 ns ± 343.383 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █ ▅ ▃ ▁                                                        
  █▂█▂█▂█▂▇▂▆▂▆▂▃▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅▂▂▁▂▅▁▂▂▄▂▂▂▄▂▂▁▃▂▁▁▃▂▂▁▃ ▃
  706 ns           Histogram: frequency by time         1.92 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.



# this PR
julia> @benchmark Bessels.airyaix_large_args(x) setup=(x=(8.3 + rand()*2)*cispi(2*rand()))
BenchmarkTools.Trial: 10000 samples with 952 evaluations.
 Range (min … max):   93.103 ns … 217.356 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):      98.878 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   107.704 ns ±  17.242 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▆ ▆█▄▂       ▃▆▁ ▆▅                   ▁▁▁                     ▂
  █▇████▇▄▃▄▄▁▃███████▄▄▃▁▃▃▁▁▃▃▄▆▆▇█▆▇▇████▆▅▃▁▄▁▄▃▄▄▄▁▁▁▅▅▇▇▇ █
  93.1 ns       Histogram: log(frequency) by time        188 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

So in this range the mean time to compute this benchmark is 9x faster. I've also set this up to be able to use SIMDMath to further reduce the time in a different PR. O and these compute both the derivative and the function value where old ones we would need two of those evaluations.