@heltonmc 2023-04-04
I added two instances of fastmath only on the single divisions to compute inverse. This doesn't affect the accuracy but has substantial effect on timing.

```julia
#no fast
julia> @benchmark Bessels.airybix_large_args(x) setup=(x=(8.3 + rand()*2)*cispi(2*rand()))
BenchmarkTools.Trial: 10000 samples with 932 evaluations.
 Range (min … max):  107.608 ns … 208.797 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     119.296 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   117.900 ns ±   4.166 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                          ▇                 █  ▁  ▂▆▇▆▂  ▂▁      
  ▂▂▂▁▁▁▁▁▁▂▃▅▃▂▁▁▂▄▇█▃▂▁▃██▃▂▂▂▂▁▂▂▁▂▂▂▁▁▁▃█▆██▃▅█████▅███▅▂▂▂ ▄
  108 ns           Histogram: frequency by time          122 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

# fast_inv
julia> @benchmark Bessels.airybix_large_args(x) setup=(x=(8.3 + rand()*2)*cispi(2*rand()))
BenchmarkTools.Trial: 10000 samples with 956 evaluations.
 Range (min … max):   90.730 ns … 204.802 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     101.634 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   101.366 ns ±  10.150 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

     ▁   ▄█▇                                                     
  ▂▂▄█▇▂▂███▄▂▂▂▂▂▂▂▂▂▂▁▂▁▂▁▁▁▁▁▂▂▁▁▁▁▂▁▁▁▂▁▁▂▁▂▁▂▁▂▁▁▁▂▁▁▁▂▁▁▂ ▃
  90.7 ns          Histogram: frequency by time          170 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
So about ~17-20 nanoseconds here.