@heltonmc 2023-04-04
I've added SIMDMath here. Need to resolve the version dependencies etc on julia and documentation. I'm fine bumping Bessels require julia v1.8. Part of me thinks we should just require v1.9 soon when that drops as precompilation is so powerful for this package.

Anway, I need to add some methods in SIMDMath - mainly methods for `fnmuladd` for complex numbers so I can switch that over to here.

But if you want to see the true awesomeness of that package....

```julia

# old benchmark
julia> @benchmark Bessels.airyaix_large_args(x) setup=(x=(8.3 + rand()*2)*cispi(2*rand()))
BenchmarkTools.Trial: 10000 samples with 952 evaluations.
 Range (min … max):   93.103 ns … 217.356 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):      98.878 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   107.704 ns ±  17.242 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▆ ▆█▄▂       ▃▆▁ ▆▅                   ▁▁▁                     ▂
  █▇████▇▄▃▄▄▁▃███████▄▄▃▁▃▃▁▁▃▃▄▆▆▇█▆▇▇████▆▅▃▁▄▁▄▃▄▄▄▁▁▁▅▅▇▇▇ █
  93.1 ns       Histogram: log(frequency) by time        188 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

# with SIMDMath
julia> @benchmark Bessels.airyaix_large_args(x) setup=(x=(8.3 + rand()*2)*cispi(2*rand()))
BenchmarkTools.Trial: 10000 samples with 985 evaluations.
 Range (min … max):  54.362 ns … 123.783 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     56.106 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   62.405 ns ±  10.390 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █▁  ▆▁                                               ▂▄▇     ▂
  ██▄▄██▆▄▇▆▃▅▃▃▄▅▄▅▃▃▅▄▃▁▃▁▄▄▁▄▁▁▃▄▁▁▃▁▁▁▃▃▁▁▁▃▁▇█▃▁▁▅████▄▅█ █
  54.4 ns       Histogram: log(frequency) by time      78.4 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

So ya - that's computing the asymptotic expansion in entire complex seconds in like 50 nanoseconds...

If you want to see the SpecialFunctions comparison of same benchmark.....

```julia
julia> import SpecialFunctions

julia> @benchmark SpecialFunctions.airyaix(x) setup=(x=(8.3 + rand()*2)*cispi(2*rand()))
BenchmarkTools.Trial: 10000 samples with 226 evaluations.
 Range (min … max):  328.677 ns …   2.421 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):       1.386 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):     1.059 μs ± 498.986 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   █▆                                                  ▃▄▂▁      
  ▄██▄▂▂▁▁▁▁▁▁▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▄▂▁▂▁▁▁▁▁▁▁▁▂▅█████▇▆▄▄ ▃
  329 ns           Histogram: frequency by time         1.54 μs <

 Memory estimate: 16 bytes, allocs estimate: 1.
```
So 25x faster....