@heltonmc 2023-04-05
I think ready to merge now. I've updated all CI and compat to julia v1.8. I think let's just plan for a big v0.3.0 release which by that time julia v1.9 should definitely have dropped.

I'll work next on improving the power series for the complex airy functions and then the intermediate range (probably going to use the sequence transformation from other thread). That should significantly improve performance in other regions.